[REFACTOR] Fix redundant tests
