--- tests/__init__.py
@@ -1,1 +1,0 @@
-"""Tests package for pantautular-be"""
